Adapted fogbow-cli with newest version of fogbow-manager
