+++ src/test/resource/ldap-identity-plugin.conf
@@ -1,0 +1,5 @@
+ldap_base =
+ldap_identity_url =
+ldap_encrypt_type = 
+private_key_path =
+public_key_path =
